Added base test case
