+++ net.powermatcher.test.osgi/src/net/powermatcher/test/osgi/BasicClusterTests.java
@@ -3,7 +3,6 @@ package net.powermatcher.test.osgi;
 import java.util.List;
 import java.util.Map;
 
-import junit.framework.TestCase;
 import net.powermatcher.api.messages.BidUpdate;
 import net.powermatcher.api.monitoring.events.IncomingPriceUpdateEvent;
 import net.powermatcher.api.monitoring.events.OutgoingBidEvent;
@@ -16,38 +15,9 @@ import net.powermatcher.examples.Freezer;
 import net.powermatcher.examples.PVPanelAgent;
 import net.powermatcher.examples.StoringObserver;
 
-import org.apache.felix.scr.ScrService;
-import org.osgi.framework.BundleContext;
-import org.osgi.framework.FrameworkUtil;
-import org.osgi.framework.ServiceReference;
 import org.osgi.service.cm.Configuration;
-import org.osgi.service.cm.ConfigurationAdmin;
 
-public class ClusterTest extends TestCase {
-
-	private final BundleContext context = FrameworkUtil.getBundle(getClass()).getBundleContext();
-    private ServiceReference<?> scrServiceReference = context.getServiceReference( ScrService.class.getName());
-    private ScrService scrService = (ScrService) context.getService(scrServiceReference);
-    private ConfigurationAdmin configAdmin;
-
-    private ClusterHelper clusterHelper;
-    
-    @Override 
-    protected void setUp() throws Exception {
-    	super.setUp();
-
-    	clusterHelper = new ClusterHelper();
-  	
-    	configAdmin = clusterHelper.getService(context, ConfigurationAdmin.class);
-
-    	// Cleanup running agents to start with clean test
-    	Configuration[] configs = configAdmin.listConfigurations(null);
-    	if (configs != null) {
-        	for (Configuration config : configs) {
-        		config.delete();
-        	}
-    	}
-    }
+public class BasicClusterTests extends OsgiTestCase {
 
     /**
      * Tests a simple buildup of a cluster in OSGI and sanity tests.
@@ -55,7 +25,7 @@ public class ClusterTest extends TestCase {
      */
     public void testSimpleClusterBuildUp() throws Exception {
     	// Create Auctioneer
-    	Configuration auctioneerConfig = clusterHelper.createConfiguration(configAdmin,clusterHelper.getFactoryPidAuctioneer(), clusterHelper.getAuctioneerProperties(clusterHelper.getAgentIdAuctioneer(), 5000));
+    	Configuration auctioneerConfig = clusterHelper.createConfiguration(configAdmin, clusterHelper.getFactoryPidAuctioneer(), clusterHelper.getAuctioneerProperties(clusterHelper.getAgentIdAuctioneer(), 5000));
 
     	// Wait for Auctioneer to become active
     	clusterHelper.checkServiceByPid(context, auctioneerConfig.getPid(), Auctioneer.class);
