+++ net.powermatcher.test.osgi/src/net/powermatcher/test/osgi/BidNumbersTests.java
@@ -2,7 +2,6 @@ package net.powermatcher.test.osgi;
 
 import java.util.List;
 
-import junit.framework.TestCase;
 import net.powermatcher.api.monitoring.events.IncomingPriceUpdateEvent;
 import net.powermatcher.core.auctioneer.Auctioneer;
 import net.powermatcher.core.concentrator.Concentrator;
@@ -10,44 +9,15 @@ import net.powermatcher.examples.Freezer;
 import net.powermatcher.examples.PVPanelAgent;
 import net.powermatcher.examples.StoringObserver;
 
-import org.apache.felix.scr.ScrService;
-import org.osgi.framework.BundleContext;
-import org.osgi.framework.FrameworkUtil;
-import org.osgi.framework.ServiceReference;
 import org.osgi.service.cm.Configuration;
-import org.osgi.service.cm.ConfigurationAdmin;
 
-public class BidNumbersTest extends TestCase {
+public class BidNumbersTests extends OsgiTestCase {
 
-	private final BundleContext context = FrameworkUtil.getBundle(getClass()).getBundleContext();
-    private ServiceReference<?> scrServiceReference = context.getServiceReference( ScrService.class.getName());
-    private ScrService scrService = (ScrService) context.getService(scrServiceReference);
-    private ConfigurationAdmin configAdmin;
-    
-    private ClusterHelper clusterHelper;
-    
-    @Override 
-    protected void setUp() throws Exception {
-    	super.setUp();
-    	
-    	clusterHelper = new ClusterHelper();
-
-    	configAdmin = clusterHelper.getService(context, ConfigurationAdmin.class);
-
-    	// Cleanup running agents to start with clean test
-    	Configuration[] configs = configAdmin.listConfigurations(null);
-    	if (configs != null) {
-        	for (Configuration config : configs) {
-        		config.delete();
-        	}
-    	}
-    }
-
-    /**
+	/**
      * Tests a simple buildup of a cluster in OSGI and sanity tests.
      * Custer consists of Auctioneer, Concentrator and 2 agents.
      */
-    public void testSimpleClusterBuildUp() throws Exception {
+    public void testVariableRateAgents() throws Exception {
     	// Create Auctioneer
     	Configuration auctioneerConfig = clusterHelper.createConfiguration(configAdmin, clusterHelper.getFactoryPidAuctioneer(), clusterHelper.getAuctioneerProperties(clusterHelper.getAgentIdAuctioneer(), 1000));
